+++ Frame128/Coin.cpp
@@ -17,7 +17,7 @@ CCoin::CCoin()
 	m_eID = OBJID::OBJ_ITEM;
 	m_eRenderID = RENDERID::RENDER_OBJ;
 
-	int i = MGR(CRandomMgr)->getFromFloatTo(1, 2);
+	int i = int(MGR(CRandomMgr)->getFromFloatTo(1, 2));
 	m_pTexInfo = MGR(CTextureMgr)->Get_Texture(L"Item", L"Coin", i);
 
 	D3DXMatrixScaling(&m_matScale, m_vScale.x / m_pTexInfo->tImgInfo.Width, m_vScale.y / m_pTexInfo->tImgInfo.Height, 0.f);
@@ -62,7 +62,7 @@ void CCoin::Render(HDC hDC)
 	MGR(CDevice)->Get_Sprite()->SetTransform(&m_matWorld);
 	MGR(CDevice)->Get_Sprite()->Draw(m_pTexInfo->pTexture, // 텍스처 객체
 		nullptr,	// 출력할 이미지 영역에 대한 렉트 구조체 포인터, null인 경우 이미지의 0, 0기준으로 출력
-		&D3DXVECTOR3(m_pTexInfo->tImgInfo.Width / 2, m_pTexInfo->tImgInfo.Height / 2, 0.f),	// 출력할 이미지의 중심 축에 대한 vec3 구조체 포인터, null인 경우 0, 0이 중심 좌표
+		&D3DXVECTOR3(float(m_pTexInfo->tImgInfo.Width / 2), float(m_pTexInfo->tImgInfo.Height / 2), 0.f),	// 출력할 이미지의 중심 축에 대한 vec3 구조체 포인터, null인 경우 0, 0이 중심 좌표
 		nullptr,	// 위치 좌표에 대한 vec3 구조체 포인터, null인 경우 스크린 상 0,0,에 좌표 출력
 		D3DCOLOR_ARGB(255, 255, 255, 255));
 	for (auto& iter : m_vecComponents)
